Correcting Test & Adding Magic Method
Summary:
Old test was failing because we were saying something like:
WeatherReport.snow_depth

and while we had the AJ1 snow depth field saved, we didn't have it accessible to the parent object. This implements a magic method to automatically return what you're looking for if there is a component object for it.

Test Plan:
Unit tests added